requirements.txt
opencv-python
ultralytics
fastapi
uvicorn
pillow
python-multipart
polars
